+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/FeaturePost.kt
@@ -38,7 +38,7 @@ data class FeaturePost (
 
     /* The post id. */
     @SerializedName("post_id")
-    val postId: kotlin.Double
+    val postId: kotlin.Int
 
 ) {
