+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/LockPost.kt
@@ -37,7 +37,7 @@ data class LockPost (
 
     /* The post id. */
     @SerializedName("post_id")
-    val postId: kotlin.Double
+    val postId: kotlin.Int
 
 ) {
